+++ lib/ash_dyan/domain/info.ex
@@ -4,12 +4,13 @@ defmodule AshDyan.Domain.Info do
   """
 
   alias AshDyan.Dsl.Domain.AnalyzableResource
+  alias Spark.Dsl.Extension
 
   @doc "Returns the list of registered analyzable resources for a domain."
   @spec analyzable_resources(module()) :: [module()]
   def analyzable_resources(domain) do
     domain
-    |> Spark.Dsl.Extension.get_entities([:dyan])
+    |> Extension.get_entities([:dyan])
     |> Enum.map(fn %AnalyzableResource{resource: resource} -> resource end)
   end
 end