+++ test/support/repo.ex
@@ -1,5 +1,11 @@
 defmodule AshDyan.Test.Repo do
   @moduledoc false
 
-  use AshPostgres.Repo, otp_app: :ash_dyan
+  use AshPostgres.Repo,
+    otp_app: :ash_dyan,
+    warn_on_missing_ash_functions?: false
+
+  def min_pg_version do
+    %Version{major: 14, minor: 0, patch: 0}
+  end
 end